+++ index.php
@@ -2,7 +2,7 @@
 /**
  * App Name:    Family Directory Management
  * Description: A simple family directory management system.
- * Version:     4.1.2
+ * Version:     4.1.4
  * 
  * @package     FamilyDirectory
  * @author      Matt Ryan